requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
Pillow
zbarlight